test for existence before deleting the file

+++ debian/sync-repo.sh
@@ -44,7 +44,7 @@ if ! [ -f $OLD_MANIFEST ]; then
 	log "Manifest file $OLD_MANIFEST not available. Aborting.\n";
 fi
 
-rm $GITLOG
+test -e $GITLOG && rm $GITLOG
 
 for PROJ in $PROJECTS;
 do
